+++ CHANGELOG.md
@@ -18,3 +18,10 @@ iOS camera onIdle delegate 수정,
 
 ## 0.8.6
 hotfix - Marker 속성중 captionTextSize, subCaptionTextSize, Alpha 의 android에서의 데이터 변환 오류 수정
+
+## 0.8.7
+- LocationOverlay Class 추가.
+    - 위치 오버레이 : 사용자의 위치를 표시하기 위해 사용되는 오버레이
+    - setPosition : LatLng 객체를 전달하여 특정 위치로 위치 오버레이 이동
+    - setBearing : double 형으로 특정 방위각을 전달하여 위치오버레이의 방위각 조정(0은 정북쪽을 의미)
+- NaverMapController 의 getSize() type 변환에러 해결